requirements.txt
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask==3.0.0
mysql-connector-python==8.2.0